Improved Menus and Custom Command Features
Also fixed a bug where menu items could be "shift clicked" out of the
inventory. The behavior is strange because it seems like a client desync
issue. In survival, the item disappears when clicked in the player
inventory. In creative, it will perform the function. This fix refreshes the inventory which stops
the desync from happening.

I'm very happy with the additions to the custom command class. The
automatic help generation will hopefully save a lot of time for devs in
the long term.