Cache Manager V2 + improved documentation
I wrote a new player caching system that is much improved from the original.
It stores data in a database instead of a YML file.
It also features in-memory caching to limit database call latency.
It finally supports looking up Bedrock player information instead of needing the Bedrock player to have joined.
Additionally, it implements fallback APIs when seeking information of players who are offline.
Best of all, it's properly async with `CompletableFuture`s

The next goal is to write a new menu system to support async items (ex. skull menu item with a texture based on a `CachedPlayer`)