Significantly improved menu system
I'm pleased to commit a new menu system that supports item refreshing.
This was necessary to display heads with textures from the async caching system, among other reasons.

It seems stable as implemented in LibreHeads. Other plugins will be updated in due time.

Note: MenuV2 is the old menu system while Menu is the new one.
This is because MenuV2 replaced my original Menu system that has since been removed.
MenuV3 felt too clunky of a name, so Menu it is.